Add passive healthcheck for consumers.
CanConnectToMessageQueueProxyHealthcheck performs a connect, a read, and
a disconnect during checkAdvanced(). Performance of the healthcheck has
been seen to be very variable and may be causing spurious alerts. The
passive healthcheck is an alternative that simply inspects the result of
the most recent "real" application read operation by its consumer
service.